feat(nomenclature-page-responsive): WIP
