+++ front/src/components/autonomous/CategoryProduct/desktop/index.ts
@@ -1,0 +1,1 @@
+export * from './CategoryProductDesktop';
