+++ front/src/components/autonomous/CategoryProduct/mobile/index.ts
@@ -1,0 +1,1 @@
+export * from './CategoryProductMobile';
