--- front/src/components/autonomous/ModalCategoryProduct/index.ts
@@ -1,1 +1,0 @@
-export * from './ModalCategoryProduct';
